+++ NEWS.md
@@ -3,6 +3,7 @@
 -   Add DOI: <https://doi.org/10.5281/zenodo.10556679>.
 -   `resmush_file()`, `resmush_dir()` and `resmush_url()` gain an new argument
     `overwrite`.
+-   `resmush_clean_dir()` shows messages with the summary of the process.
 
 # resmush 0.0.1